Implemented AI, Shove
just general work on smoothing how the battle system plays. apparently
if you put a space at the beginning of a thread it doesnt run, because
why the fuck not.

+++ GameTest/src/items/Weapon.java
@@ -10,7 +10,7 @@ public class Weapon extends Item{
 	}
 	public boolean ranged;
 	public double damage;
-	public double speed;
+	public int speed;
 	public double reach;
 	public String ammoType;
 }
